add new character route. add menu options for different new entity options
