+++ apps/game-master-client/app/components/signout.tsx
@@ -1,0 +1,11 @@
+import { Button } from "./ui/button";
+
+export function SignoutButton() {
+	return (
+		<form method="POST" action="/logout">
+			<Button variant={"outline"} size={"sm"} type="submit">
+				Logout
+			</Button>
+		</form>
+	);
+}
