+++ apps/game-master-client/app/routes/_app.games.$gameId.characters._index/route.tsx
@@ -3,10 +3,11 @@ import { typedjson, useTypedLoaderData } from "remix-typedjson";
 import { z } from "zod";
 import { parseForm, parseParams } from "zodix";
 import { api } from "~/lib/api.server";
-import { CreateCharacter } from "./components/create-character";
 import { methodNotAllowed } from "~/util/responses";
 import { validateUser } from "~/lib/auth.server";
 import { Link } from "~/components/ui/link";
+import { CreateCharacterSlideover } from "~/components/forms/create-character";
+import { createCharacterAction } from "~/queries/create-character";
 
 export const loader = async ({ request, params, context }: LoaderFunctionArgs) => {
 	const { gameId } = parseParams(params, { gameId: z.string() });
@@ -16,17 +17,9 @@ export const loader = async ({ request, params, context }: LoaderFunctionArgs) =
 	return typedjson({ gameId, gameChars });
 };
 
-export const action = async ({ request, params, context }: ActionFunctionArgs) => {
-	const userId = await validateUser(request);
+export const action = async ({ request }: ActionFunctionArgs) => {
 	if (request.method === "POST") {
-		const data = await parseForm(request, {
-			name: z.string(),
-			content: z.string(),
-			htmlContent: z.string(),
-			gameId: z.string(),
-		});
-		const result = await api.characters.createCharacter({ ...data, ownerId: userId });
-		return typedjson(result);
+    return createCharacterAction(request);
 	}
 
 	return methodNotAllowed();
@@ -36,7 +29,7 @@ export default function CharacterIndex() {
 	const { gameId, gameChars } = useTypedLoaderData<typeof loader>();
 	return (
 		<div>
-			<CreateCharacter gameId={gameId} />
+			<CreateCharacterSlideover gameId={gameId} />
 			<div className="flex flex-col gap-2 items-start">
 				{gameChars.map((char) => (
 					<Link
