+++ apps/game-master-client/app/routes/_app.games.$gameId.characters.new/route.tsx
@@ -1,7 +1,19 @@
+import type { ActionFunctionArgs } from "@remix-run/node";
+import { useParams } from "@remix-run/react";
+import { CreateCharacterForm } from "~/components/forms/create-character";
+import { createCharacterAction } from "~/queries/create-character";
+
+export const action = async ({ request }: ActionFunctionArgs) => {
+	if (request.method === "POST") {
+		return createCharacterAction(request);
+	}
+};
+
 export default function NewCharacterRoute() {
+	const params = useParams();
 	return (
 		<div>
-			<p>This is the new character route</p>
+			<CreateCharacterForm gameId={params.gameId!} />
 		</div>
 	);
 }
